Increase sleep time before startup configurations, add option in stop script for offline chatbot

+++ docker/tools/stop-bluexolo.sh
@@ -13,11 +13,17 @@
 
 #set -o nounset  # Treat unset variables as an error
 
+OPTION=$1
+
 # Kill BlueXolo and BlueXoloAssistant running containers
 docker kill bluexolo bluexolo_assistant
 
-# Remove network for offline virtual assistance
-docker network rm offline_assistant
+# Remove offline execution resources
+if [ $OPTION == "--offline" ]
+then
+    docker kill language_server
+    docker network rm offline_assistant
+fi
 
 exit 0
